+++ src/components/GroupJoin/GroupGalleryCard.story.js
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/vue'
 import { action } from '@storybook/addon-actions'
 
-import GroupsGalleryCard from './GroupsGalleryCard'
+import GroupGalleryCard from './GroupGalleryCard'
 import i18n from '@/i18n'
 
 import { groupsMock } from '>/mockdata'
@@ -11,24 +11,24 @@ const methods = {
   visit: action('visit group'),
 }
 
-storiesOf('GroupsGalleryCard', module)
+storiesOf('GroupGalleryCard', module)
   .add('isMember = true', () => ({
-    components: { GroupsGalleryCard },
-    template: '<GroupsGalleryCard :group="group" :isMember="true" @visit="visit" @preview="preview" />',
+    components: { GroupGalleryCard },
+    template: '<GroupGalleryCard :group="group" :isMember="true" @visit="visit" @preview="preview" />',
     data () { return { group: groupsMock[0] } },
     methods,
     i18n,
   }))
   .add('isMember = false', () => ({
-    components: { GroupsGalleryCard },
-    template: '<GroupsGalleryCard :group="group" :isMember="false" @preview="preview" />',
+    components: { GroupGalleryCard },
+    template: '<GroupGalleryCard :group="group" :isMember="false" @preview="preview" />',
     data () { return { group: groupsMock[0] } },
     methods,
     i18n,
   }))
   .add('without public description', () => ({
-    components: { GroupsGalleryCard },
-    template: '<GroupsGalleryCard :group="group" :isMember="false" @preview="preview" />',
+    components: { GroupGalleryCard },
+    template: '<GroupGalleryCard :group="group" :isMember="false" @preview="preview" />',
     data () {
       return {
         group: { ...groupsMock[0], publicDescription: '' },
