+++ src/components/GroupJoin/GroupGalleryCard.vue
